Have commit hooks rebuild and publish the lists for jira-projects/ADBLK#1
The process is a little contrived: `pre-commit` will let you stage files, but not add them into the current commit.

So, instead, we write to a lockfile which `post-commit` checks for. If present, it'll rebuild the lists and amend the commit to include them.

The lockfile is used because otherwise the commit --amend will re-trigger `post-commit` giving an infinite loop